+++ scripts/ourDeployRated.ts
@@ -1,5 +1,8 @@
+///// WIP
+///// WIP
+///// WIP
 import { Signer } from 'redspot/types/provider';
-import { setupOracle, setupEmmitedToken, setupCollateralMock, setupMeasurer, setupVault, setupVaultController } from './ourHelpers';
+import { setupOracle, setupEmmitedToken, setupCollateralMock, setupMeasurer, setupVault, setupVaultController } from './ourHelpersOld';
 import { consts } from './constants';
 import { fromSigner } from './helpers';
